+++ scripts/run_test.sh
@@ -1,2 +1,3 @@
-cargo t --test "*" --all-features
+cargo t --test anonymous_struct --test tuple --all-features
 cargo t --doc --all-features
+cargo +nightly miri t --all-features --test combine_futures